fix(core): r1 review — propagate inner Error, drop const rebuild, extract fixture, RegisterFn to plugin_api (refs #231)
HIGH-1: call_register now carries the inner glibre::Error enumerator name
  in ErrorContext::detail via glibre::variant_code_string() — a const char*
  pointing to a static to_string literal (plugin-abi.md §step 9 mandates
  "carrying the inner error in ErrorContext::detail").  Test extended to
  assert detail == "PluginInitFailed" for the failing-stub case.

LOW-5: drop `const` from rebuild_schedule() — future real implementation
  must mutate the schedule graph; const was a stub artefact.

LOW-6: add PROVISIONAL SIGNATURE NOTE to migrate_components declaration
  documenting that (from_version, to_version) will change shape when
  plan #221 lands the real migration table walk.

LOW-7: move RegisterFn typedef from plugin_entry.hpp to plugin_api.hpp so
  the engine loader can include plugin_api.hpp without pulling in the
  extern "C" plugin prototype declarations.  plugin_loader_registry.hpp
  now includes plugin_api.hpp instead of plugin_entry.hpp.

MED-4: extract TU-local empty-shell definitions for World/TypeRegistry/etc.
  into tests/core/plugin_loader_register/plugin_context_fixture.hpp.
  Single guarded fixture header eliminates the ODR-landmine pattern.
  Test TU updated to #include the fixture header.
